# target/classes/openapi/api-out.yaml
openapi: 3.0.1
info:
  title: ValkyrAI CORE API
  description: The API for accessing ValkyrAI services
  termsOfService: https://valkyrlabs.com/tos/
  contact: {email: info@valkyrlabs.com}
  license: {name: AGPL 3.0, url: 'https://www.gnu.org/licenses/agpl-3.0.html'}
  version: 0.9.25
...
servers:
- {url: 'http://localhost:8080/v1', description: local server (uses live data)}
- {url: 'https://valkyrlabs.com/v1', description: production server}
...
paths:
...
  /SalesActivity:
    get:
      tags: [SalesActivity]
      summary: Retrieve a list of SalesActivitys
      description: Retrieves a list of SalesActivitys.
      x-spring-cacheable: true
      operationId: getSalesActivityList
      responses:
        '200':
          description: Successfully retrieved the list of SalesActivitys.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/SalesActivity'}
        '404': {description: No SalesActivitys were found.}
        '500': {description: Internal server error.}
    post:
      tags: [SalesActivity]
      summary: Create a new SalesActivity
      description: Creates a new SalesActivity.
      operationId: postSalesActivity
      requestBody:
        description: SalesActivity details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SalesActivity'}
      responses:
        '201':
          description: SalesActivity created successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SalesActivity'}
        '400': {description: Invalid input.}
        '404': {description: SalesActivity not found.}
        '500': {description: Internal server error.}
  /SalesActivity/{id}:
    get:
      tags: [SalesActivity]
      summary: Retrieve a single SalesActivity
      description: Retrieves a single SalesActivity for a specific uid.
      x-spring-cacheable: true
      operationId: getSalesActivity
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the SalesActivity.
      responses:
        '200':
          description: Successfully retrieved the SalesActivity.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SalesActivity'}
        '404': {description: SalesActivity not found.}
        '500': {description: Internal server error.}
    put:
      tags: [SalesActivity]
      summary: Update an existing SalesActivity
      description: Updates an existing SalesActivity.
      operationId: updateSalesActivity
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the SalesActivity.
      requestBody:
        description: Updated SalesActivity details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SalesActivity'}
      responses:
        '200':
          description: SalesActivity updated successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SalesActivity'}
        '400': {description: Invalid input.}
        '404': {description: SalesActivity not found.}
        '500': {description: Internal server error.}
    delete:
      tags: [SalesActivity]
      summary: Delete a SalesActivity.
      description: Deletes a specific SalesActivity.
      operationId: deleteSalesActivity
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the SalesActivity.
      responses:
        '204': {description: SalesActivity deleted successfully.}
        '404': {description: SalesActivity not found.}
        '500': {description: Internal server error.}
...
components:
...
  schemas:
...
    SalesActivity:
      type: object
      description: Represents a sales activity or engagement related to an opportunity.
      properties:
        opportunityId: {type: string, format: uuid, description: The associated opportunity.}
        type:
          type: string
          enum: [call, meeting, email, follow-up]
          description: Type of the sales activity.
        details: {type: string, description: Details about the activity.}
        activityDate: {type: string, format: date-time, description: Date and time
            of the activity.}
        performedBy: {type: string, format: uuid, description: The user who performed
            the activity.}
        id:
          description: Unique identifier for object in the system
          type: string
          x-field-extra-annotation: "@Id  \n        @GeneratedValue(generator = \"\
            UUID\")\n"
          format: uuid
          example: d1291c3e-0260-4bc5-92e9-8931b67fc35a
        ownerId: {description: UUID of owner of the object in the system, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.CREATED_BY,
            enabled = true)', format: uuid, example: 2b9f9be4-683a-4113-ac6d-7067d94037bc}
        createdDate: {description: Date of object creation, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.CREATED_DATE, enabled = true)', format: date-time,
          example: '2025-01-03T16:21:41.598-0800'}
        keyHash: {description: 'Data, including hash of the key(s) used to encrypt
            this record.', type: string}
        lastAccessedById: {description: Last user to access object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_BY,
            enabled = true)', format: uuid, example: bba43247-148c-44b5-b60e-3eca205c491d}
        lastAccessedDate: {description: Timestamp of last access of object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.598-0800'}
        lastModifiedById: {description: Unique identifier for user who last modifed
            the object in the system, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.LAST_MODIFIED_BY, enabled = true)', format: uuid,
          example: 5600d31a-f57e-499f-8454-20c54f0eba79}
        lastModifiedDate: {description: Date of last object modification, type: string,
          x-field-extra-annotation: '    @AuditingField(fieldType = AuditingField.FieldType.LAST_MODIFIED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.598-0800'}
      required: [opportunityId, type, activityDate, performedBy]
...
  securitySchemes:
    thorapi_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://{username}.valkyrlabs.com:{port}/oauth
          scopes: {'write:items': modify owned items, 'read:items': read owned items}